EXOJCR-1862 : fixing memory leak issue.

+++ exo.jcr.component.core/src/main/java/org/exoplatform/services/jcr/impl/core/query/lucene/RedoLog.java
@@ -175,7 +175,10 @@ class RedoLog
       }
       try
       {
-         dir.deleteFile(REDO_LOG);
+         if (dir.fileExists(REDO_LOG))
+         {
+            dir.deleteFile(REDO_LOG);
+         }
       }
       catch (IOException e)
       {
